move utilities to submodule
