+++ README.md
@@ -35,21 +35,20 @@ A collection of common functions/manipulations used in Actuarial Calculations.
 - `accum_offset` to calculate accumulations like survivorship from a mortality vector
 - `spread` will calculate the spread needed between two yield curves to equate a set of cashflows
 
-### Options Pricing
-- `eurocall` and `europut` for Black-Scholes option prices (note: API may change for this in future)
-
 ### Risk Measures
 
 - Calculate risk measures for a given vector of risks:
-  - `CTE` for the Conditional Tail Expectation, or
-  - `VaR` for the percentile/Value at Risk.
+  - `CTE` for the Conditional Tail Expectation
+  - `VaR` for the percentile/Value at Risk
+  - `WangTransform` for the Wang Transformation
+  - `ProportionalHazard` for proportional hazards
+  - `DualPower` for dual power measure
 
 ### Insurance mechanics
 
 - `duration`:
   - Calculate the duration given an issue date and date (a.k.a. policy duration)
   
-
 ### Typed Rates
 
 - functions which return a rate/yield will return a `FinanceCore.Rate` object. E.g. `irr(cashflows)` will return a `Rate(0.05,Periodic(1))` instead of just a `0.05` (`float64`) to convey the compounding frequency. This is compatible across the JuliaActuary ecosystem and can be used anywhere you would otherwise use a simple floating point rate.
