+++ docs/src/utilities.md
@@ -1,0 +1,12 @@
+# Utilities submodule
+
+Provides miscellaneous routines common in actuarial and financial work.
+
+## API
+
+### Exported API
+```@docs
+Utilities.duration
+Utilities.years_between
+Utilities.accum_offset
+```
